Refactors core architecture
Migrates to an entity-component system (ECS) architecture.

Removes previous object management and input handling systems, replacing them with more flexible and scalable components and systems. This improves separation of concerns and allows for easier extension of game logic.

Adds a game mode to manage game state and adds a chunk generator for procedural level generation.

Fixes: moves from the old ControllableRectangle object to player as the new default rectangle.

This also adds support for drawing text to the screen.